Move the `packages/js/github-actions` directory under the `packages` directory
