+++ .github/scripts/github-actions-create-pr-for-release.mjs
@@ -1,10 +1,10 @@
 /**
  * Internal dependencies
  */
-import PackageTool from '../../packages/js/github-actions/utils/package-tool.js';
+import PackageTool from '../../packages/github-actions/utils/package-tool.js';
 
 export default async ( { github, context, refName, version } ) => {
-	const packageTool = new PackageTool( 'packages/js/github-actions' );
+	const packageTool = new PackageTool( 'packages/github-actions' );
 	const { heading, content } = packageTool.getChangelogByVersion( version );
 
 	const title = `Release version ${ version } of the \`github-actions\` package`;
