+++ .github/workflows/github-actions-prepare-release.yml
@@ -30,28 +30,28 @@ jobs:
         uses: actions/checkout@v3
 
       - name: Prepare node
-        uses: ./packages/js/github-actions/actions/prepare-node
+        uses: ./packages/github-actions/actions/prepare-node
         with:
           node-version: 14
-          cache-dependency-path: ./packages/js/github-actions/package-lock.json
+          cache-dependency-path: ./packages/github-actions/package-lock.json
           install-deps: "no"
 
       # The checkout revision of this repo itself doesn't have the built actions,
       # so it needs to build them before using it locally.
       - name: Build actions
         run: |
-          cd ./packages/js/github-actions
+          cd ./packages/github-actions
           npm ci --ignore-scripts
           npm run build
           cd -
 
       - name: Get release notes
         id: get-notes
-        uses: ./packages/js/github-actions/actions/get-release-notes
+        uses: ./packages/github-actions/actions/get-release-notes
         with:
           repo-token: ${{ secrets.GITHUB_TOKEN }}
-          package-dir: packages/js/github-actions
-          config-path: packages/js/github-actions/release-notes-config.yml
+          package-dir: packages/github-actions
+          config-path: packages/github-actions/release-notes-config.yml
           tag-template: "actions-v{version}"
           minor-keywords: feature, update, enhancement
 
@@ -59,7 +59,7 @@ jobs:
         # Use the github-actions bot account to commit.
         # https://api.github.com/users/github-actions%5Bbot%5D
         run: |
-          cd ./packages/js/github-actions
+          cd ./packages/github-actions
 
           TODAY=$(date '+%Y-%m-%d')
           NEXT_VER="${{ steps.get-notes.outputs.next-version }}"
