+++ README.md
@@ -7,8 +7,8 @@ The packages here are too experimental or too Grow-specific to be shared Woo-wid
 
 ## List of packages
 
+- [`/packages/github-actions`](packages/github-actions/README.md) - Custom GitHub actions.
 - [`/packages/js/generator-grow`](packages/js/generator-grow/README.md) - Yeoman Generator for extension repository boilerplate.
-- [`/packages/js/github-actions`](packages/js/github-actions/README.md) - Custom GitHub actions.
 - [`/packages/js/storybook`](packages/js/storybook/README.md) - Storybook dependencies and basic scripts
 - [`/packages/js/tracking-jsdoc`](packages/js/tracking-jsdoc/README.md) - `jsdoc` plugin to document Tracking Events in markdown
