+++ packages/github-actions/actions/stylelint-annotation/src/stylelintFormatter.js
