+++ packages/php/github-actions/readme.md
@@ -1,3 +1,3 @@
 # PHP GitHub Actions Readme
 
-View the main documentation for GitHub actions in this repository [here](/packages/js/github-actions/README.md).
+View the main documentation for GitHub actions in this repository [here](/packages/github-actions/README.md).
